app_streamlit/requirements.txt
pathlib
sys
pickle
joblib
numpy
pandas
streamlit
matplotlib.pyplot